TPASPKT-643 Preserve failed jobs for a week
Preserve failed jobs for a week instead of one year as is the default
for RQ.

In our case, the "failure TTL" means for how long a failed job can be
inspected in the job queue before it's removed. This can be useful
when using a RQ dashboard such as "rq-dashboard".