TPASPKT-643 Ensure RQ and MongoDB task status sync
The task status in MongoDB might become out-of-sync if the RQ work horse
is killed due to oom-killer or some other reason. In this case, the
failure is correctly recorded to Redis but not MongoDB.

Ensure the task status is correct in both places when retrieving
individual tasks. This prevents the `task_status` API method from
reporting a dead task as still being under processing.