TPASPKT-673 Enable ProxyFix to fix polling URL
fairdatapreservation-dev forwards requests to passipservice-dev,
requiring the use of ProxyFix to ensure that the correct hostname is
found.

Otherwise polling URL for the wrong host will be generated, leading to
403 Forbidden errors when the client tries to access the server directly
instead of going through the reverse proxy.

Earlier versions of Werkzeug trusted X-Forwarded-Host automatically,
while later versions require the developer to explicitly configure the
number of proxies. In the earlier case, a malicious party could add a
X-Forwarded-Host header early in the chain of requests, which would be
carried all the way to the final endpoint unless it was overwritten by a
proxy.