TPASPKT-720 Move response creation to API package
Code that creates HTTP responses is clearly related to API, and should
be in the api/v1 package. If another API is required, for example API
v2, the responses will be different.

Correspondingly, some code that is not related to API (for example
creating directory for temporary files, updating used quota etc.) was
moved to upload module.

Also a minor change in API: If upload request is sent without
"Content-Length" header, API response will have status code 411 ("Length
required") instead of 400 ("Bad request").