TPASPKT-820 Delete directories atomically
Delete directories atomically by first moving the directory to a
temporary location and only then updating its Metax metadata and
deleting it for real.

Each temporary directory has an unique random identifier to prevent race
conditions. Albeit unlikely, this could happen if the user deletes a
directory, creates a directory with that same name and deletes the directory
again while the first deletion is still in progress.