TPASPKT-820 Fix deleted directory after lock
If two requests attempt to delete a directory at the same time, one of
them will manage to delete it first, while the other will try to delete
the directory after acquiring the lock. In this case, the second request
will crash since renaming the disappeared directory won't work.

Handle this by returning a 404 response as with other missing files and
directories.