TPASPKT-895 Fix Metax metadata after dir delete
Metax metadata wasn't deleted after the local project directory was
deleted. This happened because the `root_upload_path` parameter was
provided to `delete_all_metadata` with the wrong depth: the path should
be provided *without* the leading project part, which was the case until
the refactoring in fdf9eaca. The paths on Metax side would be derived
from this path, which were incorrect and would result in no files
getting deleted at all.

Fix the issue by providing the parent directory instead.