TPASPKT-735 Reuse incremental MD5 checksum for DB
If the user provides a MD5 checksum to verify the file against, reuse
the same checksum when saving an entry to the `checksums` database
collection instead of reading the entire file again.

This requires a change to the API. Namely, `finalize` in
`upload_rest_api.checksum.calculate_incr_checksum` now only deletes the
Redis checkpoint, while the function always returns a checksum on each
invocation.