TPASPKT-909 Replace version.py with setuptools_scm
Replace custom script version.py with setuptools_scm, that can infer the
project version number from git tags.

Setting SETUPTOOLS_SCM_PRETEND_VERSION as a build argument is needed,
because at the moment the RPM building happens in docker containers, and
the .git folder is not copied with the project to the container. Without
the .git folder setuptools_scm doesn't have access to the version
information, but the version can be given with the
SETUPTOOLS_SCM_PRETEND_VERSION variable.