TPASPKT-936 Fix cleanup crash due to identifier
Missing identifier causes the Mongo cleanup command to crash, since the
task assumes each file has the `identifier` field. Since the same
command removes the `identifier` field for matching files, this is no
longer true when the cleanup is done for the second time.

Add check for `identifier` to fix the issue.