TPASPKT-914 Refactor database to MongoEngine
Vast majority of our database operations are basic CRUD operations,
which are better done using an ODM. Rewrite the database module to use
this library, which also provides an API similar to that used in Django.

This results in less code and also allows us to implement field
validation in some places where it wasn't done before.