TPASPKT-980 Disable synchronous archive validation
Synchronous archive validation was disabled, because in some cases
checking conflicts seems to be too slow to be done synchronously.
Archive is now validated as a background job.

Validating archive type and size are probably fast enough to be done
synchronously, so those tasks could be done synchronously, but at the
moment we do not have time to implement and test that.