TPASPKT-1023 Fix cleanup deleting non-tus uploads
Rename cleanup command from 'cleanup tus-uploads' to 'cleanup uploads'.

Set a flag for tus uploads in the database and filter the uploads that
are cleaned up to those that have the flag.

In addition, add a new 'started_at' timestamp for all uploads that is used for
cleaning non-tus uploads. Only uploads older than 2 days are cleaned; at
this point we can reasonably expect that the upload has crashed instead
of still being in progress.