TPASPKT-1265 Remove incremental checksum check
rehash is not supported for RHEL9 due to lack of support for OpenSSL 3.
As a workaround, we will be implementing checksum calculation and
verification in a background task for large files, which helps us avoid
long HTTP requests.

Remove the support for incremental checksum calculation and check
entirely. Trying to maintain the existing rehash support while
implementing the background task would complicate the implementation
even more.