TPASPKT-1265 Use background task for large upload
If the size of the uploaded file/archive exceeds a configurable
threshold (1 GB by default), perform upload finalization (checksum
calculation and verification, and moving to pre-ingest file storage) in
a background task. Individual file uploads below the size threshold
finish instantly like previously.

This is implemented as a replacement for incremental checksum
calculation that previously performed this incrementally over each HTTP
request; this meant that the last tus HTTP response could complete
almost instantly, with the exception of archive uploads that still needed
an asynchronous background job for extraction. This is done to ensure
that large file uploads do not hog WSGI workers or potentially cause
connection timeouts due to lack of activity when the server-side
checksum calculation and verification is taking place.

Previously we could assume that the tus workspace is removed once the
final tus HTTP response has been returned. This is no longer the case,
which means we have to pass the tus workspace path to the background
task which will take care of removing it.

fddps-frontend will need to be updated to track the background tasks for
uploads that do not complete instantly. As with archive uploads, the tus
identifier is used as the identifier for the corresponding background
task.

Finally, since we now run background tasks for individual files, we
could potentially leverage this for other tasks of indeterminate
length, such as generating technical metadata using file-scraper.